refactor: 지역, 원격 repository 분리
