+++ app/src/main/java/campus/tech/kakao/map/PlaceApplication.kt
@@ -4,14 +4,13 @@ import android.app.Application
 import android.net.ConnectivityManager
 import android.net.Network
 import android.net.NetworkCapabilities
-import campus.tech.kakao.map.data.PlaceDB
-import campus.tech.kakao.map.data.PlaceRepositoryImpl
+import campus.tech.kakao.map.data.PlaceRemoteDataRepository
 import campus.tech.kakao.map.domain.repository.PlaceRepository
 import com.kakao.vectormap.KakaoMapSdk
 
 class PlaceApplication: Application() {
 
-    val placeRepository: PlaceRepository by lazy { PlaceRepositoryImpl(PlaceDB(this)) }
+    val placeRepository: PlaceRepository by lazy { PlaceRemoteDataRepository(this) }
 
     override fun onCreate() {
         super.onCreate()
