+++ app/src/main/java/campus/tech/kakao/map/domain/repository/PlaceRepository.kt
@@ -7,7 +7,7 @@ interface PlaceRepository {
     suspend fun updatePlaces(places:List<Place>)
     suspend fun getPlaceById(id: String):Place?
     suspend fun getLogs(): List<Place>
-    suspend fun updateLogs(placeLog: List<Place>)
+    suspend fun updateLogs(logs: List<Place>)
     suspend fun removeLog(id: String)
     suspend fun saveLastVisitedPlace(place: Place)
     suspend fun getLastVisitedPlace(): Place?
